Update README.md
Space Station Site 
This page is a sample for  Space Travel Company. It's styled using CSS and contains a navbar, mage containers, and a form.
Movie Theater Site
This page is an example home page for a movie theater. Styled using Bootstrap 4 and contains many Bootstrap components including but not limited to a navbar, jumbotron, form, cards, and dropdowns.